+++ eir-fmm/src/reference_count/mod.rs
@@ -1,7 +1,7 @@
 mod clone;
 mod drop;
 mod pointers;
-mod utilities;
+mod records;
 
 pub use clone::*;
 pub use drop::*;
